LU-2456 lnet: Basic support for sysfs (user and kernel space)
This is an infrastructure addition to support the use of
sysfs as a part of the Dynamic LNet Config project.

Sysfs is going to be the main way to communicate
configuration changes to the LNet module.

Note: the format of sysfs comamnds is: <pid>:<command>.
The "<pid>:" is put before the reply when you read
the sysfs file.  This wil help user-space code to
watch for collisions with other processes when writing
to sysfs files (<pid> should match your own pid value).

Signed-off-by: Doug Oucharek <doug.s.oucharek@intel.com>
Change-Id: Ifc16ab6b9b7b7e98e46cda2f23ef072f45d612c5