LU-0000 osp: FOR TESTING, NOT FOR LANDING
with OSP there is a risk of getting a lot of commit callbacks.
say, 10K unlinks/sec on 4-striped files could result in 4*10K*5
= 200K commit callbacks. doesn't sound neat. so OSP introduced
a different schema where one id and commit callbacks are shared
between different OSPs. still 40K commit callbacks. in this
patch another schema is implemented: every OSP registers own
callback every second. this should result in 4*5 commit callbacks
in the same situation. in case of forced sync the commit callback
is registered unconditionally.

the patch removes th_tags from struct thandle as it is not used
with this change.

Signed-off-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>
Change-Id: I3c2f4b68df36e1ab0499a48b132042a5c5469579