LU-2139 osc: Add OBD_BRW_SOFT_SYNC page flag
This change adds a BRW page flag, OBD_BRW_SOFT_SYNC. This flag is
intended to urge the server to commit this clients unstable pages
to stable storage. Until this happens, then unstable pages are pinned
in the client's cache.

This patch only implements the client portion of the interaction. A
follow up patch will need to update the server to properly handle this
flag when it is seen, potentially syncing any unstable data.

Signed-off-by: Prakash Surya <surya1@llnl.gov>
Change-Id: I08946a5ec6afab390655643f6d3c8320bf010aea