LU-2233 tests: improve tests sanityn/40-46
sanity/40-46 usually take 800-900s which is almost a half
of the whole sanityn pass. 99.(9)% of time the tests just
wait to ensure specific order the operations execute in.

cfs_fail_intr is introduced to interrupt OBD_FAIL_TIMEOUT().
this way the tests can unblock the threads to resumt the
execution.

the tests itself are modified to use cfs_fail_intr. to be
able to do so both operations (referenced as OP1 and OP2
in the tests) are run in background. once started and then
ensured with pdo_sched() helper that MDS threads got to the
blocking points, we can interrupt OP1 and do usual checks.

on the local setup sanityn take: 1076s and 524s after.

Signed-off-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>
Change-Id: Ie01bd6a077333f6f57e533a73f38588a073a2381