LU-2188 tests: Fix assumptions in test 133d
The test assumed that with 512 files in a directory, the inode sizes
would be different in the two test directories.  This is not the case
on zfs which caused the get_rename_size() function to return multiple
values.  This change adds a argument to specify which stat is pulled
from rename_stats, and doesn't rely on the sizes being different.

Signed-off-by: Nathaniel Clark <nathaniel.l.clark@intel.com>
Change-Id: I568ec95bd7f0613caf96101055a392ea5762cd2d