LU-1289 build: clean up module versioning
Clean up module versions to match Lustre build versions using
LUSTRE_VERSION_STRING.  A few modules that are maintained
outside the lustre tree (ptllnd, ralnd, mxlnd) keep their
version as-is in case it has some meaning to the developers.
Otherwise, it would be possible to just fold the "version"
parameter into the cfs_module() macro.

Note that it is non-trivial to split out the module init
and exit into separate module_init() and module_exit()
wrappers due to the way that modules are handled in OSX.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I1c3fe5698c7f41d971a38225650597c916500c1e