LU-2432 ptlrpc: Use SLAB cache for request buffers
Instead of relying on (potentially vmalloc backed) anonymous allocations
for ptlrpc request buffers, this change modifies each ptlrpc service to
contain its own SLAB cache. Any request buffer allocation is then
allocated from, and freed back to, these named caches.

The main motivation for this patch was caused as a result of vmalloc
allocation taking a long time to complete due to a Linux kernel
regression. MDS threads would be found down in a ptlrpc_alloc_rqbd
function call, spinning on the kernel vmap_area_lock.

For example, this stack was seen on a production MDS:

    Pid: 34505, comm: mdt02_051

    Call Trace:
     [<ffffffff81273155>] ? rb_insert_color+0x125/0x160
     [<ffffffff81149f1f>] ? __vmalloc_area_node+0x5f/0x190
     [<ffffffff810609ea>] __cond_resched+0x2a/0x40
     [<ffffffff814efa60>] _cond_resched+0x30/0x40
     [<ffffffff8115fa88>] kmem_cache_alloc_node_notrace+0xa8/0x130
     [<ffffffff8115fc8b>] __kmalloc_node+0x7b/0x100
     [<ffffffffa05a2a40>] ? cfs_cpt_vmalloc+0x20/0x30 [libcfs]
     [<ffffffff81149f1f>] __vmalloc_area_node+0x5f/0x190
     [<ffffffffa05a2a40>] ? cfs_cpt_vmalloc+0x20/0x30 [libcfs]
     [<ffffffff81149eb2>] __vmalloc_node+0xa2/0xb0
     [<ffffffffa05a2a40>] ? cfs_cpt_vmalloc+0x20/0x30 [libcfs]
     [<ffffffff8114a199>] vmalloc_node+0x29/0x30
     [<ffffffffa05a2a40>] cfs_cpt_vmalloc+0x20/0x30 [libcfs]
     [<ffffffffa0922ffe>] ptlrpc_alloc_rqbd+0x13e/0x690 [ptlrpc]
     [<ffffffffa09235b5>] ptlrpc_grow_req_bufs+0x65/0x1b0 [ptlrpc]
     [<ffffffffa0927fbd>] ptlrpc_main+0xd0d/0x19f0 [ptlrpc]
     [<ffffffffa09272b0>] ? ptlrpc_main+0x0/0x19f0 [ptlrpc]
     [<ffffffff8100c14a>] child_rip+0xa/0x20
     [<ffffffffa09272b0>] ? ptlrpc_main+0x0/0x19f0 [ptlrpc]
     [<ffffffffa09272b0>] ? ptlrpc_main+0x0/0x19f0 [ptlrpc]
     [<ffffffff8100c140>] ? child_rip+0x0/0x20

In addition to this patch ensuring request buffers are not backed by
vmalloc allocation, it increases the visibility into the allocation
patterns of each ptlrpc servce. This is due to SLAB allocation
statistics being exported through the /proc/slabinfo enty in Linux.

Signed-off-by: Prakash Surya <surya1@llnl.gov>
Change-Id: I16f0fb257e3577b043cdf1bf08ee59c185539c2e