LU-2266 tests: Wait for client reconnect in recovery-small 27
In commit de311fb7 a bug was introduced where we no longer wait
for client reconnect and instead the check is such that the client
have not attempted reconnect yet. As such the test is a noop most
of the time.
Correct the waiting condiotion to include CFS_FAILED mask.

Change-Id: I3502e17184016a5c526b77a37cc7b505d8bfd5ed
Signed-off-by: Oleg Drokin <green@whamcloud.com>