LU-642 lov: prepare OSC import/export for deactivated OST
When OSC is inactivate before lov tries to connect it, lov_connect
will not connect the OST device, and the import to it is set to
invalid, when we activate it later, the obd_connect RPC is still not
issued, thus its import and export are not valid.

Since obd_connect is async RPC, fix could be enabling OSC connect
even for deactivated OST, hence its import and export are prepared.

Signed-off-by: Bobi Jam <bobijam@whamcloud.com>
Change-Id: I4aca442bc4a77baa2bc14adaf2b0f8f5c7bfa7df