LU-2266 tests: Wait for client reconnect in recovery-small 27
In commit de311fb7 a bug was introduced where we no longer wait
for client reconnect and instead the check is such that the client
have not attempted reconnect yet. As such the test is a noop most
of the time.
Correct the waiting condition to include CFS_FAILED mask.

Additinally fixed a race where MDS might reconnect to OST faster than
when we are able to set fail_loc (which is very possible on multinode
setup, original test was designed for a single node test)
(I belive this should also work on 32bit clients, since we converted to
printing unsigned longs now)

Also fix signed-unsigned issue in proc_fail_loc

Change-Id: I3502e17184016a5c526b77a37cc7b505d8bfd5ed
Signed-off-by: Oleg Drokin <green@whamcloud.com>