LU-2193 ldlm: print resource in lvbo_init() error
Print the first two lock resource fields (both the SEQ and VER/OID)
if there is an error with loading the object from disk.  This will
be more important with FID-on-OST and also the MDS.

Add the fid_extract_resid() helper, even though I ended up deciding
not to use it in the LDLM code, since I think it would be a layering
violation (the LDLM does not know what resource type it is locking).

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I7f323f15a5bf4a12865a64b2cd2bb0ebe68c6e78