LU-1718 llite: Restore NFS export for Lustre on early kernels
For kernels at 2.6.18 or earlier and kernels at 3.0+ the
struct file_system_type changed the get_sb function to a
new function called mount which was different in that the
vfsmount data was no longer passed in. The vfsmount data
was used by the llite layer for nfs export function called
get_name to search for filp that was then used with the
ll_readdir method. The approach to solve this change was
to go the route of btrfs and gfs2 to refactor some of the
llite methods to implement a directory scan independent of
filp which could be shared with nfs export funtionality.

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: I9cd8413a126ec1ea45fa5c73228e9b711e1e83a6