LU-2096 lprocfs: symlinks for forward compatibility
Currently, the "ofd" device is still named "obdfilter" in the Lustre
configuration and in /proc/fs/lustre.  It would be better to name it
"ofd" as the module is currently named.

The per-OST device statistics should really be under "ost/{target}",
like the per-MDT statistics are under "mdt/{target}" already in 2.x.

Similarly, the "OSS" statistics have always been under "ost/OSS", but
would be better in their own "oss" subdirectory.  The same is true
for "MDS" statistics, but that needs a much more complex patch.

Add symlinks in lprocfs for "oss -> ost/OSS", "ofd -> obdfilter", and
"ost/{target} -> obdfilter/{target}" for forward compatibility so that
when we finally are able to make these changes to the module names
userspace tools and scripts can have a smoother transition between
Lustre versions.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: Ia0835e39aba7c8a540a6247bd3d56bc14e500c1e