LU-2264 test:  Create memory and cpu load for testing
It can be useful to add extra memory and cpu pressure for testing.
Client nodes are likely to use extra resources.

Via processor.sh we can add cpu load.  It has a dumb bash loop
that is spaws by defaut 2 threads per cpu.  It is adjustable
as to the number of threads per cpu via the -m option.

memory_processor.c is a straight forward program that consumes
all but 100mb of total Ram and Swap. The program consumes then
exits. It should be called in a loop from a script.

This submission is to get this code out for basic review.  The
question of how to intergarete this so it is used is still
outstanding.

Test-Parameters: fortestonly
Signed-off-by: Keith Mannthey <keith.mannthey@intel.com>i
Change-Id: I590b08479c559a846fee166d2c8ea28cbe6a89e4