LU-1606 idl: remove LASSERT/CLASSERT from lustre_idl.h
Remove the usage of LASSERT() and CLASSERT() from lustre_idl.h, so
that it is usable from userspace programs if needed.  These have
crept in over the years, but are not intended to be there.

The CLASSERT() checks for fid swabbing were largely redundant, and
have been moved to lustre/fid/fid_hander.c.

The checks for lustre_lma_init() was moved to local_object_create()
and got a comment to also update lustre_lma_swab().

The fid_is_sane() macro has been modified to remove the check that
f_ver == 0, since this will eventually not be true, and the client
or server should not crash if it sees such a FID during usage.

There are still a few LASSERTs that need to be removed when FID-on-OST
is landed, but I don't want to remove them before that code lands.

There are also uses of CERROR() that could be removed at that time.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I04901d7713dc31b489685d4639e7bf857decab07