LU-1199 build: Fix style and indention in build system files
Fix indentation by replacing spaces with tabs.

Remove stray whitespace at end of lines.

Command substitution changed from `...` notation to $(...) notation.
The latter is part of the POSIX 1003.2, and is supported by Debian's
dash shell, so it is relativly portable these days.

build/Makefile is checked in to git, not auto-generated by the
build system.  Therefore a !Makefile exception is added to
build/.gitignore.

Also, in build/autogen.sh, fix default setting of the IFS variable
to be a space and a tab, rather than 9 spaces.  I believe the 9
spaces were a result of a bad cut-and-paste.

Change-Id: I46270d808d7fe036f748da1cb45826bb42ab01e2
Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>