LU-1866 dt: update interface for dt_index_operations
Although it can be implemented indirectly by "remove + insert",
the direct "update" interface is more efficient and semanteme
clean, and avoid a lot of lock handling for contention.

The dt_index_operations "update" interface is mainly used for
updating the record in index file, and for some test purposes.

Currently, the main user is LFSCK and OI scrub.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I4ca08e8bd6617fcadd63d7803e55ea6e0f9a1ae2