LU-2492 obdclass: htable_lookup could miss a waking up signal
In lu_object_put(), cfs_hash_bd_dec_and_lock() could possibly count
the refcount added by htable_lookup(), so if the thread run to this
code path is the last object holder, it should have gone down to the
free part, but because the htable_lookup() messes the refcount, the
last object holder misses the change to free the object,
table_lookup() will then always find the 0 referred object in the hash
table, which no one will delete it from hash table and free it any
longer, and lu_object_find_at() keeps waiting forever.

In this case, we'd do the unhash and free process in htable_lookup().

Signed-off-by: Bobi Jam <bobijam@gmail.com>
Change-Id: I4bd57e944bf9167d70d97845be55241b538872df