LU-2498 mount: Set 'deadline' scheduler on ldiskfs
The recommended IO scheduler to use with Lustre on top of ldiskfs is the
'deadline' scheduler. This patch attempts to automatically make this
configuration change for the user at mount time, by setting the IO
scheduler to 'deadline' on the device. This makes it unnecessary for the
user to use any out of band mechanism to ensure this value is set
properly (i.e. start up scripts, kernel configuration, etc.).

Signed-off-by: Prakash Surya <surya1@llnl.gov>
Change-Id: If869468ba5370afddb29233dd893d74b9c45f367