LU-2483 build: Update Copyrights for b2_1
In prepration of the release of 2.1.4 the Copyrights need to be
updated.

As I have bound the script to the changes so that it will not be lost
from the changes it helped make.

The accounting works for both Intel and Whamcloud email addresses.
See: lustre/ldlm/ldlm_request.c
It is able to add new Copyrights when aproprate
See: lustre/ptlrpc/layout.c
I didn't find any Intel, Inc signed files.
Two man pages were edited by hand and are included in this commit.

Signed-off-by: Keith Mannthey <keith.mannthey@intel.com>
Change-Id: I0c6daac2c6bfd2b6d5ed1f14bb5993519d099b91